Move custom gtk-based objects to a separate module
Next step is to lazy-load the modules.
ve
